requirements.txt
# Core ML Libraries
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0

# Experiment Tracking
mlflow==2.7.1

# API Framework
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6

# Data Versioning
dvc==3.22.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
matplotlib==3.7.2
seaborn==0.12.2
